Moved class definitions out of /vent/io/devices/__init__.py and into /vent/io/devices/base.py
